geo add

+++ Gallify/Models/Playlist/Playlist.swift
@@ -75,8 +75,8 @@ class Playlist: Encodable, Decodable, ObservableObject, Identifiable {
     @Published var playlistId: String
     @Published var description: String
 
-    @Published var lat: Int //large decimal.
-    @Published var lon: Int
+    @Published var lat: Double //large decimal.
+    @Published var lon: Double
     @Published var address: String
     @Published var modifiedDate: Int
     @Published var createdDate: Int
@@ -204,8 +204,8 @@ class Playlist: Encodable, Decodable, ObservableObject, Identifiable {
         searchType = try container.decode(String.self, forKey: .searchType)
         playlistId = try container.decode(String.self, forKey: .playlistId)
         description = try container.decode(String.self, forKey: .description)
-        lat = try container.decode(Int.self, forKey: .lat)
-        lon = try container.decode(Int.self, forKey: .lon)
+        lat = try container.decode(Double.self, forKey: .lat)
+        lon = try container.decode(Double.self, forKey: .lon)
         address = try container.decode(String.self, forKey: .address)
         modifiedDate = try container.decode(Int.self, forKey: .modifiedDate)
         createdDate = try container.decode(Int.self, forKey: .createdDate)
